+++ coordinator/src/notifications.rs
@@ -1,5 +1,4 @@
-use crate::db::positions;
-use crate::db::user;
+use crate::db::positions_helper::get_positions_joined_with_fcm_token_with_expiry_within;
 use crate::position::models::Position;
 use anyhow::Context;
 use anyhow::Result;
@@ -21,6 +20,7 @@ const END_OF_EXPIRED_POSITION: time::Duration = time::Duration::hours(1);
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum NotificationKind {
+    RolloverWindowOpen,
     PositionSoonToExpire,
     PositionExpired,
 }
@@ -30,6 +30,7 @@ impl Display for NotificationKind {
         match self {
             NotificationKind::PositionSoonToExpire => write!(f, "PositionSoonToExpire"),
             NotificationKind::PositionExpired => write!(f, "PositionExpired"),
+            NotificationKind::RolloverWindowOpen => write!(f, "RolloverWindowOpen"),
         }
     }
 }
@@ -115,6 +116,10 @@ fn build_notification<'a>(kind: NotificationKind) -> fcm::Notification<'a> {
             notification_builder.title("Your position has expired");
             notification_builder.body("Open the app to react.");
         }
+        NotificationKind::RolloverWindowOpen => {
+            notification_builder.title("Rollover window is open");
+            notification_builder.body("Open the app to rollover your position for the next cycle.");
+        }
     }
     notification_builder.finalize()
 }
@@ -164,32 +169,11 @@ pub async fn query_and_send_position_notifications(
     conn: &mut PgConnection,
     notification_sender: &mpsc::Sender<Notification>,
 ) -> Result<()> {
-    let users = user::all(conn)?;
-
-    let positions_with_fcm_tokens = positions::Position::get_all_positions_with_expiry_within(
+    let positions_with_fcm_tokens = get_positions_joined_with_fcm_token_with_expiry_within(
         conn,
         OffsetDateTime::now_utc() - START_OF_EXPIRING_POSITION,
         OffsetDateTime::now_utc() + END_OF_EXPIRED_POSITION,
-    )?
-    .into_iter()
-    // Join positions with users to add the FCM tokens.
-    // Filter out positions that don't have a FCM token stored in the users
-    // table which is with them.
-    // This can be done at the DB level if it ever becomes a performance issue.
-    .filter_map(|p| {
-        let maybe_fcm_token = users
-            .iter()
-            .find(|u| u.pubkey == p.trader.to_string())
-            .map(|u| FcmToken(u.fcm_token.clone()));
-
-        if let Some(fcm_token) = maybe_fcm_token {
-            Some((p, fcm_token))
-        } else {
-            tracing::warn!(?p, "No FCM token for position");
-            None
-        }
-    })
-    .collect::<Vec<_>>();
+    )?;
 
     send_expiry_notifications_if_applicable(&positions_with_fcm_tokens, notification_sender).await;
 
@@ -228,6 +212,30 @@ async fn send_expiry_notifications_if_applicable(
                 .await
             {
                 tracing::error!("Failed to send PositionSoonToExpire notification: {:?}", e);
+            }
+        }
+    }
+}
+
+/// Send notifications to users with positions that they can rollover
+pub async fn send_rollover_reminder(
+    positions: &[(Position, FcmToken)],
+    notification_sender: &mpsc::Sender<Notification>,
+) {
+    let now = OffsetDateTime::now_utc();
+
+    tracing::info!("Sending reminder");
+
+    for (position, fcm_token) in positions {
+        if position.expiry_timestamp > now {
+            if let Err(e) = notification_sender
+                .send(Notification::new(
+                    fcm_token.clone(),
+                    NotificationKind::RolloverWindowOpen,
+                ))
+                .await
+            {
+                tracing::error!("Failed to send RolloverWindowOpen notification: {:?}", e);
             }
         }
     }